Refactores locations data
