+++ src/data/campusValla/rooms/B.js
@@ -1,0 +1,92 @@
+import { locations } from '../../locationsData'
+
+export default [
+  {
+    text: 'BL32',
+    id: 'b93cec2c-6da0-4b69-8cf8-a925a8832427'
+  },
+  {
+    text: 'BL33',
+    id: 'fd86f739-d763-440f-a0da-278ca5df39ae'
+  },
+  {
+    text: 'BL33',
+    id: '7d1561b3-9245-434e-ada5-ddde80ba5a4c'
+  },
+  {
+    text: 'Ada Lovelace',
+    id: 'ae945d57-2c9c-4622-9d4e-2c2a205587ce'
+  },
+  {
+    text: 'CYD-poolen',
+    id: 'cf31dc1a-bc98-4af1-92f6-fa3f737a787a'
+  },
+  {
+    text: 'ISYtan',
+    id: '61f8b52d-4c93-4217-81c3-3774566d36bb'
+  },
+  {
+    text: 'SU00',
+    id: '20a98cbc-4551-4888-b503-eb447e3e850f'
+  },
+  {
+    text: 'SU01',
+    id: '6120e166-b099-4ffc-88ef-855c44b66974'
+  },
+  {
+    text: 'SU02',
+    id: '1cfdc6fc-cedb-4137-90fe-4b9a692934d6'
+  },
+  {
+    text: 'SU03',
+    id: '45db0d85-6009-462f-a10e-fd247b48944e'
+  },
+  {
+    text: 'SU10',
+    id: '5ca53466-ba57-4318-b986-aea662a5f896'
+  },
+  {
+    text: 'SU11',
+    id: '6b68ad92-ea03-4081-8aad-adde06e440ad'
+  },
+  {
+    text: 'SU12',
+    id: 'adbdcc05-10f3-4ef4-bdee-a0a5ec3794e6'
+  },
+  {
+    text: 'SU13',
+    id: '31e7befa-e113-4dee-a8a1-0521c930af3f'
+  },
+  {
+    text: 'SU14',
+    id: '8551a5d7-4e93-4029-81b2-330bb6ce0df4'
+  },
+  {
+    text: 'SU15',
+    id: 'd29ac8af-e2d4-49e7-98f4-492227bf2265'
+  },
+  {
+    text: 'SU16',
+    id: '773bf91c-53ac-480b-8ae9-0d832fd5c64c'
+  },
+  {
+    text: 'SU17',
+    id: '1f30bf90-c456-4985-bacd-f289bea8e8b7'
+  },
+  {
+    text: 'SU18',
+    id: 'b2704e2b-b160-411d-a22d-183743211bba'
+  },
+  {
+    text: 'Svedberg',
+    id: '68e264e5-cae9-46c2-ab0b-9c363dda0528'
+  },
+  {
+    text: 'Theorell',
+    id: '7e05e180-2090-4952-82f6-6d1467baf09d'
+  }
+].map(({ text, id }) => ({
+  text,
+  id,
+  locationId: locations.campusValla['B-huset'].id
+}))
