+++ src/data/campusValla/rooms/I.js
@@ -1,0 +1,68 @@
+import { locations } from '../../locationsData'
+
+export default [
+  {
+    text: 'I101',
+    id: '0237b71d-559b-4f14-ba61-521466a69b4f'
+  },
+  {
+    text: 'I205',
+    id: '6402a119-72a8-48d9-bb89-22ce44572ff8'
+  },
+  {
+    text: 'I206',
+    id: '4e8f8b40-6282-4da5-bd73-ddbab271f545'
+  },
+  {
+    text: 'I102',
+    id: 'd4b646f4-2b10-46d9-9410-cc8564d14be9'
+  },
+  {
+    text: 'IG10',
+    id: '0a1f508f-b4bc-4ef0-b330-43b22c08f583'
+  },
+  {
+    text: 'IG11',
+    id: '3597629d-d984-47a4-8bb0-59cc7e6e2d84'
+  },
+  {
+    text: 'IG12',
+    id: '0d502e8c-0e5f-4f19-ab9b-346a26235bb2'
+  },
+  {
+    text: 'IG13',
+    id: 'e6ae403e-b00c-43cf-ae69-a5e557963999'
+  },
+  {
+    text: 'IG14',
+    id: '070c3a08-602f-4818-8ac4-4316eade50ff'
+  },
+  {
+    text: 'IG15',
+    id: '4212d75d-bf1e-4036-8db1-9b5cce6ac7e7'
+  },
+  {
+    text: 'IG16',
+    id: '93d00218-eff0-4072-85b8-d76182472002'
+  },
+  {
+    text: 'IG17',
+    id: '8823cdad-5995-402c-be2f-ee589c1a9234'
+  },
+  {
+    text: 'IG20',
+    id: '162cd498-fcec-4b96-8684-06a4c3b40e78'
+  },
+  {
+    text: 'IG21',
+    id: '3f1a50c0-ce82-4b06-a738-f516f026d087'
+  },
+  {
+    text: 'I-fiket',
+    id: '03dd5bc2-a10c-4db8-9e9f-52f72a0fb5c9'
+  }
+].map(({ text, id }) => ({
+  text,
+  id,
+  locationId: locations.campusValla['I-huset'].id
+}))
