+++ src/data/otherPlaces/outsideCampus.js
@@ -1,0 +1,16 @@
+import { locations } from '../locationsData'
+
+export default [
+  {
+    text: 'Övrigt',
+    id: 'ead531c3-372c-4dd6-9280-2c9db28ddc38'
+  },
+  {
+    text: 'SkåLand',
+    id: '352c3dde-3767-4940-aa63-a4eac42449bf'
+  }
+].map(({ text, id }) => ({
+  text,
+  id,
+  locationId: locations.others['Utanför campus'].id
+}))
