--- src/utils/campusUS.js
@@ -1,417 +1,0 @@
-export const locationsUS = {
-  'Hus 001': {
-    id: '949636dc-997a-4aa2-8582-7c81a12dad4d',
-    text: 'Hus 001'
-  },
-  'Hus 420': {
-    id: '52392466-b327-49cd-83e7-fbc1edd13f56',
-    text: 'Hus 420'
-  },
-  'Hus 421': {
-    id: '8d31f888-99c9-4caf-a3ba-c31619a658f6',
-    text: 'Hus 421'
-  },
-  'Hus 463': {
-    id: '0d8d31ba-f773-4db2-9119-0fdcbf342819',
-    text: 'Hus 463'
-  },
-  'Hus 462': {
-    id: 'f536fe7a-2bbb-4321-8edb-92346f3f5ffc',
-    text: 'Hus 462'
-  },
-  'Hus 448': {
-    id: 'be6df087-6884-4f80-87a1-8d0547ab2feb',
-    text: 'Hus 448'
-  },
-  'Hus 473': {
-    id: 'a8e79460-581b-423a-a0f5-07d1ae75b940',
-    text: 'Hus 473'
-  },
-  'Hus 240': {
-    id: 'c1173d8c-3b1e-4e91-bf5a-38fa30f23bb7',
-    text: 'Hus 240'
-  },
-  'Hus 511': {
-    id: 'faceed67-e485-4ed8-8aeb-264d72e3be52',
-    text: 'Hus 511'
-  }
-}
-
-const house001 = [
-  {
-    text: 'Humle',
-    id: '1482d707-e5d1-44ae-a8f7-3da8522a1bb2',
-    locationId: locationsUS['Hus 001'].id
-  },
-  {
-    text: 'Fläder',
-    id: 'cc229189-849c-419a-8cc7-e2375d9413ec',
-    locationId: locationsUS['Hus 001'].id
-  },
-  {
-    text: 'Rudbeckia',
-    id: 'ec6114d0-f0a2-4578-b207-4d20d7777811',
-    locationId: locationsUS['Hus 001'].id
-  },
-  {
-    text: 'Johannesört',
-    id: 'd0667434-f29c-4583-a60f-c42dd3d2c702',
-    locationId: locationsUS['Hus 001'].id
-  },
-  {
-    text: 'Bolmört',
-    id: '5b020a12-2931-4349-a58c-5d8d4945b948',
-    locationId: locationsUS['Hus 001'].id
-  },
-  {
-    text: 'Malört',
-    id: 'd7be54da-f6e2-4d4a-a87a-249d39b33db4',
-    locationId: locationsUS['Hus 001'].id
-  },
-  {
-    text: 'Vallört',
-    id: '2df5a95f-7c2b-4b42-8748-8c836cbc56fc',
-    locationId: locationsUS['Hus 001'].id
-  },
-  {
-    text: 'Kamomill',
-    id: '6c448dd9-ef2e-48f8-bc0e-13b6a742dad8',
-    locationId: locationsUS['Hus 001'].id
-  }
-]
-
-const house420 = [
-  {
-    text: 'Bunsen',
-    id: 'bb432d58-f3ee-40d3-aa8d-fd28c4299945',
-    locationId: locationsUS['Hus 420'].id
-  },
-  {
-    text: 'Mätglaset',
-    id: '6f281671-33da-44b1-9c94-3455b3b2877e',
-    locationId: locationsUS['Hus 420'].id
-  },
-  {
-    text: 'Bägaren',
-    id: '70523c60-b250-4421-9304-7d8e308b9de2',
-    locationId: locationsUS['Hus 420'].id
-  },
-  {
-    text: 'Lågan',
-    id: '98deb62c-05ab-4900-aac9-e046497bc836',
-    locationId: locationsUS['Hus 420'].id
-  },
-  {
-    text: 'Pillret',
-    id: 'ee9e58e7-ecdd-4ae0-bd67-5dc26ca2df53',
-    locationId: locationsUS['Hus 420'].id
-  },
-  {
-    text: 'Öglan',
-    id: 'ea4bf742-e341-48ae-934e-415186cf9479',
-    locationId: locationsUS['Hus 420'].id
-  },
-  {
-    text: 'Vågen',
-    id: '65bd44ea-22cd-4b26-9058-15b37b1919b8',
-    locationId: locationsUS['Hus 420'].id
-  },
-  {
-    text: 'Kolven',
-    id: 'e07d3fed-c067-42e7-adf5-7b50cf6bab0a',
-    locationId: locationsUS['Hus 420'].id
-  }
-]
-
-const house240 = [
-  {
-    text: 'Birgittasalen, HS',
-    id: '00161871-691c-420e-8929-68abef974c9b',
-    locationId: locationsUS['Hus 240'].id
-  },
-  {
-    text: 'Fasanen',
-    id: '47f748fc-9fe1-423e-a515-584f327af0c5',
-    locationId: locationsUS['Hus 240'].id
-  },
-  {
-    text: 'Ejdern',
-    id: 'b19fef21-eafe-4d1a-886b-dc901fe65bad',
-    locationId: locationsUS['Hus 240'].id
-  },
-  {
-    text: 'Tranan',
-    id: '493da18b-07b4-462e-a236-ddd7fd49c67d',
-    locationId: locationsUS['Hus 240'].id
-  },
-  {
-    text: 'Staren',
-    id: '08d7346c-8d8d-4150-b851-509b1cf20484',
-    locationId: locationsUS['Hus 240'].id
-  },
-  {
-    text: 'Lärkan',
-    id: '73b0b569-15f2-4430-9616-c72f26ef0fa6',
-    locationId: locationsUS['Hus 240'].id
-  },
-  {
-    text: 'Tjädern',
-    id: 'e0d922af-3af9-46c7-b370-a94b3e480284',
-    locationId: locationsUS['Hus 240'].id
-  },
-  {
-    text: 'Svalan',
-    id: '3afc1b67-4b85-4499-ae22-d4331b4dd2f7',
-    locationId: locationsUS['Hus 240'].id
-  },
-  {
-    text: 'Orren',
-    id: 'ef210c04-0e5e-4719-9f4a-a955cff112b4',
-    locationId: locationsUS['Hus 240'].id
-  }
-]
-
-const house421 = [
-  {
-    text: 'Eken, HS',
-    id: '1df1bbea-a0b3-4fe8-9897-a0a2579d0e33',
-    locationId: locationsUS['Hus 421'].id
-  },
-  {
-    text: 'Linden, HS',
-    id: '047e146f-400e-410f-bacd-cdddd51752e4',
-    locationId: locationsUS['Hus 421'].id
-  },
-  {
-    text: 'Björken',
-    id: 'e2d1da6e-c090-4d6e-a818-e4bdf34a217f',
-    locationId: locationsUS['Hus 421'].id
-  },
-  {
-    text: 'Almen',
-    id: 'a11b8b9c-5936-4221-b0a7-77706e6446aa',
-    locationId: locationsUS['Hus 421'].id
-  }
-]
-
-const house448 = [
-  {
-    text: 'Hugo Theorell, HS ',
-    id: '443b701c-0d58-48bc-9cb6-9ac6044513da',
-    locationId: locationsUS['Hus 448'].id
-  },
-  {
-    text: 'Granitsalen, CA',
-    id: 'aab7c369-e0b4-4f46-a195-688c95d9da5d',
-    locationId: locationsUS['Hus 448'].id
-  },
-  {
-    text: 'Antracit, ALC',
-    id: 'd09eb425-1b85-40fa-88c6-e91bde4642c8',
-    locationId: locationsUS['Hus 448'].id
-  },
-  {
-    text: 'Dolomit, IKT',
-    id: '0c153552-a46e-4180-b83c-57aecc7fae3e',
-    locationId: locationsUS['Hus 448'].id
-  },
-  {
-    text: 'Flinta',
-    id: '7a6685f3-9589-424c-bd2d-c714379c2ce8',
-    locationId: locationsUS['Hus 448'].id
-  },
-  {
-    text: 'Skiffer',
-    id: 'f2cb2b8b-aa96-4d65-942e-b591b038785c',
-    locationId: locationsUS['Hus 448'].id
-  },
-  {
-    text: 'Gnejs',
-    id: '22e6da92-4480-48ac-955b-d871099cef41',
-    locationId: locationsUS['Hus 448'].id
-  },
-  {
-    text: 'Glimmer',
-    id: 'b9c290a7-eeca-49e9-acb8-feef808c730a',
-    locationId: locationsUS['Hus 448'].id
-  },
-  {
-    text: 'Marmor',
-    id: '4fc6d69d-2c40-4426-ac5c-2a98ceb49e5c',
-    locationId: locationsUS['Hus 448'].id
-  }
-]
-
-const house462 = [
-  {
-    text: 'Nils Holger, HS',
-    id: '5f792ac2-cdd0-49fa-8ba5-9c5a039f19f6',
-    locationId: locationsUS['Hus 462'].id
-  },
-  {
-    text: 'Flöjten',
-    id: '95e67ad5-a942-4fe3-94c6-30d3c697af36',
-    locationId: locationsUS['Hus 462'].id
-  },
-  {
-    text: 'Trumpeten',
-    id: '6bdac943-5a8e-467f-82f4-2e2b5e73d571',
-    locationId: locationsUS['Hus 462'].id
-  },
-  {
-    text: 'Valthornet',
-    id: 'f1ed938a-b633-4333-a6aa-4e9997bceca2',
-    locationId: locationsUS['Hus 462'].id
-  },
-  {
-    text: 'Klarinetten',
-    id: '57084b7d-90ed-4725-8fce-e680f00a69d3',
-    locationId: locationsUS['Hus 462'].id
-  },
-  {
-    text: 'Trombonen',
-    id: 'd009a911-0129-4294-8295-4f7656f4844e',
-    locationId: locationsUS['Hus 462'].id
-  }
-]
-
-const house463 = [
-  {
-    text: 'Berzelius, HS',
-    id: '65828199-c651-4599-8f20-876d6066ecf6',
-    locationId: locationsUS['Hus 463'].id
-  },
-  {
-    text: 'Rönnen',
-    id: '0fb3ddd6-814e-4f35-87ec-8c36a99484cf',
-    locationId: locationsUS['Hus 463'].id
-  },
-  {
-    text: 'Kotten',
-    id: 'e6cad65f-703d-4cd5-a134-cd8983f5ad42',
-    locationId: locationsUS['Hus 463'].id
-  },
-  {
-    text: 'Fröet',
-    id: '4dec7a69-2d50-480f-894e-8c75a8b7080e',
-    locationId: locationsUS['Hus 463'].id
-  },
-  {
-    text: 'Barret',
-    id: 'dc02689a-c722-41ce-a794-6f8632899610',
-    locationId: locationsUS['Hus 463'].id
-  },
-  {
-    text: 'Ingrid Marie',
-    id: '5b6d121a-7658-4259-be90-c7fdbaa5e646',
-    locationId: locationsUS['Hus 463'].id
-  },
-  {
-    text: 'Cox Pomona',
-    id: '51d60e75-5aa4-4de1-9c39-78a265558c33',
-    locationId: locationsUS['Hus 463'].id
-  },
-  {
-    text: 'Aroma',
-    id: 'ab27e405-cf73-4e7d-a117-dd320da76aa3',
-    locationId: locationsUS['Hus 463'].id
-  },
-  {
-    text: 'Astrakan',
-    id: 'f120e772-d604-41d9-89a0-c11cbc13ba72',
-    locationId: locationsUS['Hus 463'].id
-  },
-  {
-    text: 'Åkerö',
-    id: 'a7dff6de-31f0-466e-9281-a768761db851',
-    locationId: locationsUS['Hus 463'].id
-  }
-]
-
-const house473 = [
-  {
-    text: 'Karl Johan',
-    id: '6d4c358f-7ff0-4e73-93e7-3746103f3968',
-    locationId: locationsUS['Hus 473'].id
-  },
-  {
-    text: 'Kantarellen',
-    id: '949a1e25-06e8-490e-9b51-d93e22fb9065',
-    locationId: locationsUS['Hus 473'].id
-  },
-  {
-    text: 'Fårtickan',
-    id: 'b8f5ff66-5d6f-42d9-8da0-bfffe9b5fce9',
-    locationId: locationsUS['Hus 473'].id
-  },
-  {
-    text: 'Murklan',
-    id: '43a6b707-88e1-40ce-9e0a-27d5f7768baa',
-    locationId: locationsUS['Hus 473'].id
-  }
-]
-
-const house511 = [
-  {
-    text: 'Hasselquist, HS',
-    id: 'd5a3c022-2394-48b5-8f89-fac6a48bf419',
-    locationId: locationsUS['Hus 511'].id
-  },
-  {
-    text: 'Belladonna, IKT',
-    id: 'a4ff7010-5ae9-46ec-87f9-a5fa22bc0892',
-    locationId: locationsUS['Hus 511'].id
-  },
-  {
-    text: 'Valeriana',
-    id: '19641b16-c840-4b43-bf78-46b97a26b182',
-    locationId: locationsUS['Hus 511'].id
-  },
-  {
-    text: 'Digitalis',
-    id: '85d55638-4a96-4ddb-be6e-ad3ac5d44396',
-    locationId: locationsUS['Hus 511'].id
-  },
-  {
-    text: 'Papaver, IKT',
-    id: 'ccb3b1ed-613d-4079-a42b-5101929427e7',
-    locationId: locationsUS['Hus 511'].id
-  },
-  {
-    text: 'Salix',
-    id: 'f7bebe9c-c885-426f-a9ab-7e6c2a695db1',
-    locationId: locationsUS['Hus 511'].id
-  },
-  {
-    text: '• Timjan',
-    id: '1ff6ed16-d11d-430d-a22b-e82416e0a8b6',
-    locationId: locationsUS['Hus 511'].id
-  },
-  {
-    text: 'Salvia',
-    id: 'd0197f41-2b11-4f6d-9587-bd913ade2fa3',
-    locationId: locationsUS['Hus 511'].id
-  },
-  {
-    text: 'Mynta',
-    id: '6ad2673c-7c75-409f-ab6c-8d15a073ec67',
-    locationId: locationsUS['Hus 511'].id
-  },
-  {
-    text: 'Kanel',
-    id: '587057c8-2a56-4b72-a793-35d702c04ec3',
-    locationId: locationsUS['Hus 511'].id
-  }
-]
-
-export const roomsUS = [
-  ...house001,
-  ...house240,
-  ...house420,
-  ...house421,
-  ...house448,
-  ...house462,
-  ...house463,
-  ...house473,
-  ...house511
-]
